requirements.txt
# FinNavigator AI - LangChain Deep Agents
# Requirements for the multi-agent financial intelligence system

# Core LangChain
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-experimental>=0.3.0

# LangChain Integrations
langchain-openai>=0.2.0
langchain-anthropic>=0.2.0

# Vector Stores
langchain-chroma>=0.1.0
chromadb>=0.5.0

# Embeddings
langchain-huggingface>=0.1.0
sentence-transformers>=3.0.0

# Web Framework
streamlit>=1.30.0
plotly>=5.18.0

# Data Processing
pandas>=2.2.3
numpy>=1.26.4

# SEC API
sec-api>=1.0.0

# LLM Clients
openai>=1.30.0

# Utilities
python-dotenv>=1.0.0
httpx>=0.28.1
requests>=2.32.3

# Memory & Storage
redis>=5.0.0

# Async
aiohttp>=3.11.16
asyncio-throttle>=1.0.2

# Scientific Computing
scipy>=1.12.0
scikit-learn>=1.4.0

# Testing & Monitoring
langsmith>=0.1.0

# Local SLM + LangGraph
langchain-ollama>=0.2.0
langgraph>=0.2.0
